abstract out plotKMeansScoringHistory
