+++ src/h2oModelOutput/createOutput.js
@@ -11,6 +11,7 @@ import downloadPojo from './downloadPojo';
 import downloadMojo from './downloadMojo';
 import exportModel from './exportModel';
 import deleteModel from './deleteModel';
+import plotKMeansScoringHistory from './plotKMeansScoringHistory';
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
@@ -108,29 +109,8 @@ export default function createOutput(_) {
   switch (_.model.algo) {
     case 'kmeans':
       table = _.inspect('output - Scoring History', _.model);
-      if (table) {
-        const gFunction = g => g(
-            g.path(
-              g.position('iteration', 'within_cluster_sum_of_squares'),
-              g.strokeColor(
-                g.value('#1f77b4')
-              )
-            ),
-            g.point(
-              g.position('iteration', 'within_cluster_sum_of_squares'),
-              g.strokeColor(
-                g.value('#1f77b4')
-              )
-            ),
-            g.from(table)
-          );
-        const plotFunction = _.plot(gFunction);
-        renderPlot(
-          _,
-          'Scoring History',
-          false,
-          plotFunction
-        );
+      if (typeof table !== 'undefined') {
+        plotKMeansScoringHistory(_, table);
       }
       break;
     case 'glm':
